Expand API to encompass image transfer. Images are now detected from question text, discovered in directory path, transferred to client who receives them in to a QList.

+++ Quiz/client.h
@@ -41,7 +41,7 @@ signals:
 private:
     QString _username;
     bool    _connected;
-    qint16  _blockSize;
+    quint32 _blockSize;
     bool    _loggedin;
     QuizModel* _model;
     bool    _sending;
